ConstructorOrderInfo does not need to be nested
Also, ObjectBuilder can be private again.
